feat(theme/dashboard): extract Posts table into RecentlyCreatedPosts component
