+++ packages/theme/src/pages/dashboard/Index.vue
@@ -4,49 +4,7 @@
   <div class="flex items-start gap-x-4 px-3 lg:px-6 py-6">
     <div class="flex-2">
       <div class="text-neutral-500 font-medium mb-2 text-sm ml-1.5">Posts</div>
-      <Table>
-        <template #header>
-          <Td
-            :head="true"
-            :style="{
-              width: '200px',
-            }"
-            class="flex-1 break-all"
-          >
-            Title
-          </Td>
-          <Td :head="true">
-            Votes
-          </Td>
-        </template>
-
-        <Tr
-          v-for="post in posts"
-          :key="post.postId"
-        >
-          <div class="relative flex items-center">
-            <Td
-              :style="{
-                width: '200px',
-              }"
-              class="flex-1 line-clamp-1 truncate hover:line-clamp-none hover:break-all hover:whitespace-normal"
-            >
-              {{ post.title }}
-            </Td>
-            <Td>
-              {{ post.voters.votesCount }}
-            </Td>
-            <router-link
-              :to="`/dashboard/posts/${encodeURIComponent(post.slug)}`"
-              class="absolute inset-0"
-            />
-          </div>
-        </Tr>
-
-        <template #infinite-loader>
-          <infinite-scroll :on-infinite="getRecentPosts" :state="postState" />
-        </template>
-      </Table>
+      <recently-created-posts />
     </div>
 
     <div class="flex-1">
@@ -57,52 +15,19 @@
 </template>
 
 <script setup lang="ts">
-import { defineAsyncComponent, onMounted, ref } from "vue";
+import { defineAsyncComponent, onMounted } from "vue";
 import { useHead } from "@vueuse/head";
 import { useRoute, useRouter } from "vue-router";
-import type { IPost } from "@logchimp/types";
-
-// modules
-import { Posts } from "../../modules/posts";
-import { useSettingStore } from "../../store/settings";
 
 // components
-import InfiniteScroll, {
-  type InfiniteScrollStateType,
-} from "../../components/ui/InfiniteScroll.vue";
-import Table from "../../components/ui/Table/Table.vue";
 import DashboardPageHeader from "../../components/dashboard/PageHeader.vue";
-import Tr from "../../components/ui/Table/Tr.vue";
-import Td from "../../components/ui/Table/Td.vue";
 
+const RecentlyCreatedPosts = defineAsyncComponent(
+  () => import("../../components/dashboard/index/RecentlyCreatedPosts.vue"),
+);
 const RecentlyCreatedBoards = defineAsyncComponent(
   () => import("../../ee/components/dashboard/index/RecentlyCreatedBoards.vue"),
 );
-const posts = ref<IPost[]>([]);
-const postState = ref<InfiniteScrollStateType>("IDLE");
-const settingsStore = useSettingStore();
-
-async function getRecentPosts() {
-  postState.value = "LOADING";
-
-  const postsAPI = new Posts();
-
-  try {
-    const response = await postsAPI.GetPosts(
-      {},
-      {
-        first: "4",
-        created: "DESC",
-      },
-    );
-
-    posts.value = response.posts;
-    postState.value = "COMPLETED";
-  } catch (error) {
-    console.error(error);
-    postState.value = "ERROR";
-  }
-}
 
 const route = useRoute();
 const routerInstance = useRouter();
